Add data mutation, message factory helpers
