+++ test/messages/UserMessage.test.js
@@ -5,6 +5,10 @@
 const test = require('ava');
 const chai = require('chai');
 
+const DataHelper = require('../helpers/DataHelper');
+const MessageFactoryHelper = require('../helpers/MessageFactoryHelper');
+
+
 // ------- Init ----------------------------------------------------------------
 
 chai.should();
@@ -12,7 +16,7 @@ chai.should();
 // ------- Tests ---------------------------------------------------------------
 
 test(() => {
-
+  DataHelper.testMessageFailsValidationWithout('id', MessageFactoryHelper.getValidUser);
 });
 
 // ------- End -----------------------------------------------------------------
